Make localStorage work for your current level and board
One problem I ran into was needing to update the board from
StorageHandler whenever the level id changed. The problem was that if I
used an effect the effect would be triggered when the level index was
changed OR whenever the board was changed. I needed it only to trigger
when the index was changed, so I had to resort to using a store instead
of an effect.